requirements.txt
Flask
opencv-python
torch
torchvision
paddleocr
numpy
Pillow
requests
mysql-connector-python